Stop Start song feature added

+++ app/src/main/java/com/example/musicplayer/model/Song.kt
@@ -1,3 +1,3 @@
 package com.example.musicplayer.model
 
-data class Song(val id: Int, val title: String, val duration: Double, val path: String)
+data class Song(val id: Int, val title: String, val duration: Long, val path: String)
